requirements.txt
flask
flask-session
google-generativeai
PyPDF2
python-docx
fpdf